hello-world/requirements.txt
black>=25.9.0
isort>=7.0.0
langchain>=1.0.3
langchain-anthropic>=1.0.1
langchain-gemini>=0.1.1
langchain-google-genai>=3.0.0
langchain-ollama>=1.0.0
langchain-openai>=1.0.1
python-dotenv>=1.2.1
playwright
playwright-stealth
humanization-playwright